Make the propram more object-oriented, to make sharing data easier
This will also pave the way for reading a configuration file, whereupon
more config data will be set, not just one `Args` object.

Also: video size is determined by the maximum dimension, not the
minimum. Because a 1920x500 video is still high resolution and will be
displayed with small pixels, regardless of the skinny height.